fix(component): dont redifine element if it exists
fix(component): move registry patch to its own module

this needs to be ran before any of the other mwc code to patch the registry.

temp fix for #162

fix(component): dont redifine element if it exists

this needs to be ran before any of the other mwc code to patch the registry.

temp fix for #162